Dev2.3.0.0

+++ Figaro/src/main/scala/com/cra/figaro/algorithm/factored/VariableElimination.scala
@@ -139,12 +139,12 @@ trait VariableElimination[T] extends FactoredAlgorithm[T] with OneTime {
   def eliminationOrder(factors: Traversable[Factor[T]], toPreserve: Traversable[Variable[_]]): List[Variable[_]] = {
     val eliminableVars = (Set[Variable[_]]() /: factors)(_ ++ _.variables) -- toPreserve
     var initialGraph = new VEGraph(factors)
-    val candidates = new HeapPriorityMap[Variable[_], Long]
+    val candidates = new HeapPriorityMap[Variable[_], Double]
     eliminableVars foreach (v => candidates += v -> initialGraph.score(v))
     eliminationOrderHelper(candidates, toPreserve, initialGraph, List())
   }
 
-  @tailrec private def eliminationOrderHelper(candidates: PriorityMap[Variable[_], Long],
+  @tailrec private def eliminationOrderHelper(candidates: PriorityMap[Variable[_], Double],
     toPreserve: Traversable[Variable[_]],
     graph: VEGraph,
     accum: List[Variable[_]]): List[Variable[_]] = {
